+++ DiscImageChef.Filesystems/FATX/Info.cs
@@ -97,13 +97,13 @@ namespace DiscImageChef.Filesystems.FATX
             {
                 Type = "FATX filesystem",
                 ClusterSize =
-                    (int)(fatxSb.sectorsPerCluster * logicalSectorsPerPhysicalSectors *
-                          imagePlugin.Info.SectorSize),
+                    (uint)(fatxSb.sectorsPerCluster * logicalSectorsPerPhysicalSectors *
+                           imagePlugin.Info.SectorSize),
                 VolumeName   = volumeLabel,
                 VolumeSerial = $"{fatxSb.id:X8}"
             };
-            XmlFsType.Clusters = (long)((partition.End - partition.Start + 1) * imagePlugin.Info.SectorSize /
-                                        (ulong)XmlFsType.ClusterSize);
+            XmlFsType.Clusters = (partition.End - partition.Start + 1) * imagePlugin.Info.SectorSize /
+                                 XmlFsType.ClusterSize;
         }
     }
 }
