+++ DiscImageChef.Filesystems/PCEngine.cs
@@ -66,7 +66,7 @@ namespace DiscImageChef.Filesystems
             XmlFsType = new FileSystemType
             {
                 Type        = "PC Engine filesystem",
-                Clusters    = (long)((partition.End - partition.Start + 1) / imagePlugin.Info.SectorSize * 2048),
+                Clusters    = (partition.End - partition.Start + 1) / imagePlugin.Info.SectorSize * 2048,
                 ClusterSize = 2048
             };
         }
